Create interfaces file and import necessary interfaces to components
