+++ src/Components/Continents/Continents.tsx
@@ -3,18 +3,7 @@ import { SyntheticEvent } from 'react'
 import { CountriesData } from '../../countries.model'
 import './Continents.css'
 import { Route, Routes, NavLink, Link } from 'react-router-dom'
-
-
-interface CountriesProps {
-  continents: CountriesData[]
-  continent?: CountriesData[]
-  assignSelections?: selections | any
-  filterSelections: (categoryData: string) => void
-}
-
-interface selections {
-  assignSelections: (newSelection: object | string) => void
-}
+import { CountriesProps} from '../../interfaces'
 
 const Continents: React.FC<CountriesProps> = (props): JSX.Element => {
   const [selectedContinent, setSelectedContinent] = useState({})
