+++ src/Components/Scoreboard/Scoreboard.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { KeepScore, Guesses, Score } from '../App/App'
+import { Guesses, Score, KeepScore} from '../../interfaces'
 
 export default function Scoreboard({ keepScore, guesses }: {keepScore: KeepScore, guesses: Guesses}) {
   const score: Score = keepScore(guesses)
